backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "logfire>=4.4.0",
    "pydantic-ai>=1.0.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
